+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 **⚠ Breaking changes**
 - `allowFootnote` becomes `allowReference`
 - `onFootnote` becomes `onReference`
+- `Element.tagName` returns the tag name in uppercase.
 
 
 ## 1.0.0 - 2020-08-13